+++ copi.owasp.org/priv/repo/migrations/20210522184517_create_cards.exs
@@ -17,6 +17,7 @@ defmodule Copi.Repo.Migrations.CreateCards do
       add :owasp_appsensor, {:array, :string}
       add :capec, {:array, :string}
       add :safecode, {:array, :string}
+      add :biml, :string, default: ""
       add :url, :string, default: ""
       timestamps()
     end