+++ copi.owasp.org/priv/repo/migrations/20250609100558_populate_mlsec_cards.exs
@@ -1,4 +1,4 @@
-defmodule Copi.Repo.Migrations.PopulateNewCards do
+defmodule Copi.Repo.Migrations.PopulateNewMlSecCards do
   use Ecto.Migration
   alias Copi.Repo
   alias Copi.Cornucopia